Redesigned method of page crawling
Complete redesign of the page crawling method. It's been designed with the intent of implementing threading on top of it, which will be added later. This redesign made it possible to implement a quick index option, which results in much less data being collected, but puts way less strain on the webserver (making everyone happier!).