Changed db storage location
Changed it to the locations provided by the module appdirs, which should allow for easy compatability between linux, windows and mac. also changed the name of the database table for the indexes in preparation for more tables in the future.

Tests will need to be cleaned up, since they are poluting the "production" db on my system.