Added CommandMenu for input parsing
This is the first of many steps for cleaning up the way user input is handled. For starters, I'm implementing a class I wrote the other day to cleanly represent a large menu of commands. Command registration now uses this menu, and the ParseInput class uses this for command validation. I've also moved a couple functions from utils to __init__ since they were only ever used once within the init.