Finished unit tests
Improved the functionality of the Tag commands. the associations tabe was renamed at one point, so I've removed that from the test database being used. Also implemented a consistent way of interacting with the database for unit tests, which should prevent database polution. It hasn't been implemented in all unit tests yet, but it's a start...